docbook pdf: add xsl parameter to locate images
Unlike html, pdf/ps requires a full path name to find images,
but only needs it to make an internal copy of it at build time.
The image can later be removed and the image still shows up in the pdf doc.
This allows us to use the absolute builddir.

This parameter must not be used for html which loads the image from disk
at user read time. The image is removed from builddir after build is done,
and we do not know where the image will be installed by the distro.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>